requirements.txt
PyYAML
requests
markdown
beautifulsoup4